+++ app/actions/index.ts
@@ -1,1 +1,3 @@
 export { addProjectMemberAction } from './addProjectMemberAction';
+export { addMetaDataStandardsAction } from './addMetaDataStandardsAction';
+export { addRepositoryAction } from './addRepositoryAction';