+++ components/QuestionAdd/actions/index.ts
@@ -1,2 +1,0 @@
-export { addRepositoryAction } from './addRepositoryAction';
-export { addMetaDataStandardsAction } from './addMetaDataStandardsAction';